Moved api to different file and added small readme detail
